Renaming functions and files for CX playground.
